refactor: move outlook_meeting and outlook_query_events to outlook-calendar plugin
