+++ plugins/outlook-mail/src/handlers.ts
@@ -392,238 +392,3 @@ export async function flagMessage(
   if (res.error) return { error: JSON.stringify(res.error) };
   return { ok: true, flag_status: params.flag_status };
 }
-
-export async function createMeeting(
-  config: OutlookMailConfig,
-  params: {
-    to: string | string[];
-    cc?: string[];
-    subject: string;
-    start: string;
-    duration?: string;
-    location?: string;
-    description?: string;
-    timezone?: string;
-    signature?: string;
-  },
-): Promise<unknown> {
-  const { clientId, clientSecret, refreshToken } = config;
-  if (!clientId) return { error: "OUTLOOK_CLIENT_ID not set" };
-  const token = await getToken(clientId, clientSecret, refreshToken, "Calendars.ReadWrite");
-
-  const startDt = new Date(params.start);
-  if (isNaN(startDt.getTime())) return { error: `Invalid start: '${params.start}'` };
-  const mins = parseDurationMins(params.duration ?? "1h");
-  const endDt = new Date(startDt.getTime() + mins * 60_000);
-  const tz = params.timezone ?? "America/Los_Angeles";
-
-  const toList = Array.isArray(params.to) ? params.to : [params.to];
-  const ccList = params.cc ?? [];
-  const allAttendees = [
-    ...toList.map(e => ({ emailAddress: { address: e }, type: "required" })),
-    ...ccList.map(e => ({ emailAddress: { address: e }, type: "optional" })),
-  ];
-
-  const eventObj: Record<string, unknown> = {
-    subject: params.subject,
-    start: { dateTime: startDt.toISOString(), timeZone: tz },
-    end: { dateTime: endDt.toISOString(), timeZone: tz },
-    attendees: allAttendees,
-    isOnlineMeeting: false,
-    ...(params.location ? { location: { displayName: params.location } } : {}),
-    ...(params.description ? { body: { contentType: "Text", content: params.description } } : {}),
-  };
-
-  const created = await graphPost(token, "/me/events", eventObj) as Record<string, unknown>;
-  if (created.error) return { error: JSON.stringify(created.error) };
-
-  return {
-    ok: true,
-    id: created.id,
-    iCalUId: created.iCalUId,
-    subject: created.subject,
-    start: (created.start as Record<string, string>)?.dateTime,
-    end: (created.end as Record<string, string>)?.dateTime,
-    webLink: created.webLink,
-    message: `✓ Meeting created: ${params.subject} (${startDt.toISOString().slice(0, 16)})`,
-  };
-}
-
-export async function updateEvent(
-  config: OutlookMailConfig,
-  params: {
-    event_id?: string;
-    find?: string;
-    new_title?: string;
-    new_start?: string;
-    new_duration?: string;
-    new_location?: string;
-    new_description?: string;
-    timezone?: string;
-    status?: string;
-    add_attendee?: string[];
-    remove_attendee?: string[];
-    force?: boolean;
-  },
-): Promise<unknown> {
-  const { clientId, clientSecret, refreshToken } = config;
-  if (!clientId) return { error: "OUTLOOK_CLIENT_ID not set" };
-  const token = await getToken(clientId, clientSecret, refreshToken, "Calendars.ReadWrite");
-
-  if (!params.event_id && !params.find) return { error: "Provide event_id or find to identify the event." };
-
-  let eventId: string;
-  let currentEvent: Record<string, unknown>;
-
-  if (params.event_id) {
-    eventId = params.event_id;
-    currentEvent = await graphGet(token, `/me/events/${encodeURIComponent(eventId)}`) as Record<string, unknown>;
-    if (currentEvent.error) return { error: JSON.stringify(currentEvent.error) };
-  } else {
-    const needle = esc(String(params.find));
-    const res = await graphGet(token, `/me/events?$filter=contains(subject,'${needle}')&$top=10&$select=id,subject,start,end,attendees`) as { value: Array<Record<string, unknown>> };
-    const matches = res.value ?? [];
-    if (matches.length === 0) return { error: `No event found matching: '${params.find}'` };
-    if (matches.length > 1 && !params.force) {
-      return {
-        error: `Found ${matches.length} matching events. Use event_id to target one, or force=true to update all.`,
-        events: matches.map(e => ({ id: e.id, subject: e.subject, start: (e.start as Record<string, string>)?.dateTime })),
-      };
-    }
-    const targets = params.force ? matches : [matches[0]];
-    const results: unknown[] = [];
-    for (const ev of targets) {
-      const r = await updateEvent(config, { ...params, event_id: String(ev.id), find: undefined });
-      results.push(r);
-    }
-    return { updated: results.length, results };
-  }
-
-  const tz = params.timezone ?? "America/Los_Angeles";
-  const patch: Record<string, unknown> = {};
-
-  if (params.new_title) patch.subject = params.new_title;
-  if (params.new_description) patch.body = { contentType: "Text", content: params.new_description };
-  if (params.new_location) patch.location = { displayName: params.new_location };
-  if (params.status) patch.showAs = params.status;
-
-  if (params.new_start) {
-    const newStart = new Date(params.new_start);
-    if (isNaN(newStart.getTime())) return { error: `Invalid new_start: '${params.new_start}'` };
-    patch.start = { dateTime: newStart.toISOString(), timeZone: tz };
-    if (params.new_duration) {
-      const mins = parseDurationMins(params.new_duration);
-      patch.end = { dateTime: new Date(newStart.getTime() + mins * 60_000).toISOString(), timeZone: tz };
-    }
-  } else if (params.new_duration) {
-    const existingStart = (currentEvent.start as Record<string, string>)?.dateTime;
-    if (existingStart) {
-      const base = new Date(existingStart);
-      const mins = parseDurationMins(params.new_duration);
-      patch.end = { dateTime: new Date(base.getTime() + mins * 60_000).toISOString(), timeZone: tz };
-    }
-  }
-
-  if (params.add_attendee?.length || params.remove_attendee?.length) {
-    const existing = (currentEvent.attendees as Array<Record<string, unknown>> | undefined) ?? [];
-    let attendees = existing.map(a => ({ ...(a as object) })) as Array<Record<string, unknown>>;
-    if (params.remove_attendee) {
-      const toRemove = new Set(params.remove_attendee.map(e => e.toLowerCase()));
-      attendees = attendees.filter(a => {
-        const email = String(((a.emailAddress as Record<string, string> | undefined)?.address) ?? "").toLowerCase();
-        return !toRemove.has(email);
-      });
-    }
-    if (params.add_attendee) {
-      for (const email of params.add_attendee) {
-        attendees.push({ emailAddress: { address: email }, type: "required" });
-      }
-    }
-    patch.attendees = attendees;
-  }
-
-  if (Object.keys(patch).length === 0) return { error: "No changes specified." };
-
-  const updated = await graphPatch(token, `/me/events/${encodeURIComponent(eventId)}`, patch) as Record<string, unknown>;
-  if (updated.error) return { error: JSON.stringify(updated.error) };
-
-  return { ok: true, id: eventId, message: `✓ Event updated: ${updated.subject ?? eventId}` };
-}
-
-export async function queryEvents(
-  config: OutlookMailConfig,
-  params: {
-    after?: string;
-    before?: string;
-    text?: string;
-    attendee?: string;
-    uid?: string;
-  },
-): Promise<unknown> {
-  const { clientId, clientSecret, refreshToken } = config;
-  if (!clientId) return { error: "OUTLOOK_CLIENT_ID not set" };
-  const token = await getToken(clientId, clientSecret, refreshToken, "Calendars.Read");
-
-  if (params.uid) {
-    const res = await graphGet(token, `/me/events?$filter=iCalUId eq '${esc(params.uid)}'`) as { value: Array<Record<string, unknown>> };
-    return { events: (res.value ?? []).map(formatEvent) };
-  }
-
-  const filters: string[] = [];
-  if (params.after) filters.push(`start/dateTime ge '${params.after}T00:00:00'`);
-  if (params.before) filters.push(`end/dateTime le '${params.before}T23:59:59'`);
-  if (params.text) filters.push(`contains(subject,'${esc(params.text)}')`);
-
-  const qs = filters.length
-    ? `?$filter=${encodeURIComponent(filters.join(" and "))}&$top=50&$orderby=start/dateTime`
-    : `?$top=20&$orderby=start/dateTime`;
-
-  const res = await graphGet(token, `/me/events${qs}`) as { value: Array<Record<string, unknown>> };
-  let events = (res.value ?? []).map(formatEvent);
-
-  if (params.attendee) {
-    const att = params.attendee.toLowerCase();
-    events = events.filter(e =>
-      (e.attendees as Array<Record<string, string>>)?.some(
-        (a: Record<string, string>) => a.email?.toLowerCase() === att,
-      ),
-    );
-  }
-
-  return { count: events.length, events };
-}
-
-// ---------------------------------------------------------------------------
-// Private helpers for calendar tools
-// ---------------------------------------------------------------------------
-
-function parseDurationMins(d: string): number {
-  const s = d.toLowerCase().trim();
-  if (s.endsWith("h")) return Math.round(parseFloat(s.slice(0, -1)) * 60);
-  if (s.endsWith("m")) return parseInt(s.slice(0, -1), 10);
-  const n = parseInt(s, 10);
-  if (isNaN(n)) throw new Error(`Invalid duration: '${d}' (use e.g. '1h', '30m', '1.5h')`);
-  return n;
-}
-
-function formatEvent(e: Record<string, unknown>): Record<string, unknown> {
-  const start = (e.start as Record<string, string> | undefined)?.dateTime ?? "";
-  const end = (e.end as Record<string, string> | undefined)?.dateTime ?? "";
-  const attendees = ((e.attendees as Array<Record<string, unknown>>) ?? []).map(a => ({
-    email: String(((a.emailAddress as Record<string, string> | undefined)?.address) ?? ""),
-    name: String(((a.emailAddress as Record<string, string> | undefined)?.name) ?? ""),
-    type: String(a.type ?? ""),
-    status: String(((a.status as Record<string, string> | undefined)?.response) ?? ""),
-  }));
-  return {
-    id: e.id,
-    iCalUId: e.iCalUId,
-    subject: e.subject,
-    start: start.slice(0, 16),
-    end: end.slice(0, 16),
-    location: (e.location as Record<string, string> | undefined)?.displayName ?? "",
-    organizer: ((e.organizer as Record<string, Record<string, string>> | undefined)?.emailAddress?.address) ?? "",
-    attendees,
-    webLink: e.webLink,
-  };
-}
