+++ plugins/outlook-mail/src/index.ts
@@ -6,7 +6,7 @@
 
 import { definePlugin } from "carapace-plugin-sdk";
 import { Type } from "@sinclair/typebox";
-import { getInbox, searchMail, readMessage, saveAttachments, sendMessage, replyToMessage, forwardMessage, moveMessage, flagMessage, createMeeting, updateEvent, queryEvents, type OutlookMailConfig } from "./handlers.js";
+import { getInbox, searchMail, readMessage, saveAttachments, sendMessage, replyToMessage, forwardMessage, moveMessage, flagMessage, type OutlookMailConfig } from "./handlers.js";
 
 export const createEntry = definePlugin({
   id: "outlook-mail",
@@ -252,91 +252,6 @@ export const createEntry = definePlugin({
             message_id: String(message_id ?? ""),
             flag_status: flag_status as "flagged" | "complete" | "notFlagged",
           });
-        } catch (e) {
-          return { error: (e as Error).message };
-        }
-      },
-    }),
-    tool({
-      name: "outlook_meeting",
-      label: "Outlook Create Meeting",
-      description: "Create a calendar meeting invite via Microsoft Graph and send invitations to attendees.",
-      parameters: Type.Object({
-        to: Type.Union([Type.String(), Type.Array(Type.String())], { description: "Attendee email address(es)." }),
-        cc: Type.Optional(Type.Array(Type.String(), { description: "Optional attendees (will be marked as optional)." })),
-        subject: Type.String({ description: "Meeting title." }),
-        start: Type.String({ description: "Start datetime in ISO format (e.g. 2026-03-15T14:00)." }),
-        duration: Type.Optional(Type.String({ description: "Duration: '1h', '30m', '1.5h' (default: 1h)." })),
-        location: Type.Optional(Type.String({ description: "Meeting location." })),
-        description: Type.Optional(Type.String({ description: "Meeting description / agenda." })),
-        timezone: Type.Optional(Type.String({ description: "IANA timezone (default: America/Los_Angeles)." })),
-        signature: Type.Optional(Type.String({ description: "Signature block for the invite email." })),
-      }),
-      async execute({ to, cc, subject, start, duration, location, description, timezone, signature }, config) {
-        try {
-          const resolvedConfig: OutlookMailConfig = {
-            clientId: String(config.clientId ?? process.env.OUTLOOK_CLIENT_ID ?? ""),
-            clientSecret: String(config.clientSecret ?? process.env.OUTLOOK_CLIENT_SECRET ?? ""),
-            refreshToken: String(config.refreshToken ?? process.env.OUTLOOK_REFRESH_TOKEN ?? ""),
-          };
-          return await createMeeting(resolvedConfig, { to, cc, subject, start, duration, location, description, timezone, signature });
-        } catch (e) {
-          return { error: (e as Error).message };
-        }
-      },
-    }),
-
-    tool({
-      name: "outlook_update_event",
-      label: "Outlook Update Event",
-      description: "Find a calendar event by ID or text search and update its title, time, location, attendees, or status.",
-      parameters: Type.Object({
-        event_id: Type.Optional(Type.String({ description: "Microsoft Graph event ID to target directly." })),
-        find: Type.Optional(Type.String({ description: "Free-text search across event title." })),
-        new_title: Type.Optional(Type.String({ description: "Replace the event title." })),
-        new_start: Type.Optional(Type.String({ description: "New start time (ISO format)." })),
-        new_duration: Type.Optional(Type.String({ description: "New duration (e.g. '1h', '30m')." })),
-        new_location: Type.Optional(Type.String({ description: "Replace location." })),
-        new_description: Type.Optional(Type.String({ description: "Replace description/notes." })),
-        timezone: Type.Optional(Type.String({ description: "Timezone for new_start (default: America/Los_Angeles)." })),
-        status: Type.Optional(Type.String({ description: "Update showAs status (free, busy, tentative, oof, workingElsewhere)." })),
-        add_attendee: Type.Optional(Type.Array(Type.String(), { description: "Email(s) to add as attendees." })),
-        remove_attendee: Type.Optional(Type.Array(Type.String(), { description: "Email(s) to remove from attendees." })),
-        force: Type.Optional(Type.Boolean({ description: "Update all matching events when multiple found." })),
-      }),
-      async execute({ event_id, find, new_title, new_start, new_duration, new_location, new_description, timezone, status, add_attendee, remove_attendee, force }, config) {
-        try {
-          const resolvedConfig: OutlookMailConfig = {
-            clientId: String(config.clientId ?? process.env.OUTLOOK_CLIENT_ID ?? ""),
-            clientSecret: String(config.clientSecret ?? process.env.OUTLOOK_CLIENT_SECRET ?? ""),
-            refreshToken: String(config.refreshToken ?? process.env.OUTLOOK_REFRESH_TOKEN ?? ""),
-          };
-          return await updateEvent(resolvedConfig, { event_id, find, new_title, new_start, new_duration, new_location, new_description, timezone, status, add_attendee, remove_attendee, force });
-        } catch (e) {
-          return { error: (e as Error).message };
-        }
-      },
-    }),
-
-    tool({
-      name: "outlook_query_events",
-      label: "Outlook Query Events",
-      description: "Query calendar events by date range, text, attendee email, or iCalUId.",
-      parameters: Type.Object({
-        after: Type.Optional(Type.String({ description: "Only events starting at or after this date (ISO, e.g. 2026-03-01)." })),
-        before: Type.Optional(Type.String({ description: "Only events starting before this date (ISO, e.g. 2026-04-01)." })),
-        text: Type.Optional(Type.String({ description: "Filter by text match on title." })),
-        attendee: Type.Optional(Type.String({ description: "Filter to events including this attendee email." })),
-        uid: Type.Optional(Type.String({ description: "Return the single event with this exact iCalUId." })),
-      }),
-      async execute({ after, before, text, attendee, uid }, config) {
-        try {
-          const resolvedConfig: OutlookMailConfig = {
-            clientId: String(config.clientId ?? process.env.OUTLOOK_CLIENT_ID ?? ""),
-            clientSecret: String(config.clientSecret ?? process.env.OUTLOOK_CLIENT_SECRET ?? ""),
-            refreshToken: String(config.refreshToken ?? process.env.OUTLOOK_REFRESH_TOKEN ?? ""),
-          };
-          return await queryEvents(resolvedConfig, { after, before, text, attendee, uid });
         } catch (e) {
           return { error: (e as Error).message };
         }
