+++ plugins/outlook-mail/tests/index.test.ts
@@ -38,9 +38,9 @@ beforeEach(() => {
 
 describe("plugin entry", () => {
   it("has correct id and name", async () => { const { entry } = await loadPlugin(); expect(entry.id).toBe("outlook-mail"); });
-  it("registers all 12 tools", async () => {
+  it("registers all 9 tools", async () => {
     const { api } = await loadPlugin();
-    expect(Object.keys(api.tools).sort()).toEqual(["outlook_flag","outlook_forward","outlook_inbox","outlook_meeting","outlook_move","outlook_query_events","outlook_read","outlook_reply","outlook_save_attachments","outlook_search","outlook_send","outlook_update_event"]);
+    expect(Object.keys(api.tools).sort()).toEqual(["outlook_flag","outlook_forward","outlook_inbox","outlook_move","outlook_read","outlook_reply","outlook_save_attachments","outlook_search","outlook_send"]);
   });
 });
